Wrapped everything in one component, displaying pages works and other small bugs fixed
