+++ app/styles/app.sass
@@ -67,9 +67,14 @@ a
 
 .editHover a
   visibility: hidden
-  
+
 .editHover:hover a
    visibility: visible
+
+#page-numbers
+  color: black
+  font-size: 30px
+  font-family: Comic Sans MS
 
 #pageButtons, #pageInputs
   display: flex
